json: JSMN_PARENT_LINKS makes the parser a billion times faster

+++ defs.h
@@ -4,6 +4,8 @@
 
 #define ARRAY_SIZE(x) (sizeof(x) / sizeof((x)[0]))
 
+#include <inttypes.h>
+
 typedef unsigned char u8;
 typedef signed char s8;
 
@@ -13,8 +15,8 @@ typedef signed short s16;
 typedef unsigned int u32;
 typedef signed int s32;
 
-typedef unsigned long long u64;
-typedef signed long long s64;
+typedef uint64_t u64;
+typedef int64_t s64;
 
 #endif /* LNVIS_DEFS_H */
